Task 0 completion

+++ main.h
@@ -14,8 +14,8 @@ int _isalpha(int c);
 int rot13(char *c);
 
 int base_o_b(unsigned long int num, int base);
-int hex(long unsigned int num, char x);
+int hex(unsigned long int num, char x);
 int print_integer(int num);
-int print_unsigned(unsigned int num);
+int print_unsigned(unsigned long int num);
 int print_addr(void *ptr);
 #endif
